requirements.txt
scipy
numpy
pandas
matplotlib.pyplot
neurokit2